Registering ConfigStore in constructor instead of StoreConfig
